refactor: rename common.monitoring package to common.logging
두 모듈의 common/monitoring은 전부 로그 관련(LogSanitizer,
LogStatusProvider, LogTestController)이고 실제 모니터링(metrics/trace)
아님. api/common/logging이 이미 존재해서 그쪽으로 통합.
